+++ .eslintrc
@@ -1,7 +1,17 @@
 {
-  "extends": ["airbnb-base", "plugin:jasmine/recommended", "prettier"],
-  "plugins": ["jasmine", "prettier"],
+  "extends": [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/eslint-recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:jasmine/recommended",
+    "prettier"
+  ],
+  "parser": "@typescript-eslint/parser",
+  "plugins": ["jasmine", "@typescript-eslint", "prettier"],
   "rules": {
+    "@typescript-eslint/no-namespace": "off",
+    "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/ban-ts-comment": "off",
     "prettier/prettier": [
       "error",
       {
@@ -9,12 +19,11 @@
         "trailingComma": "none",
         "printWidth": 100,
         "semi": true,
-        "parser": "babel"
+        "parser": "typescript"
       }
     ],
     "jasmine/no-spec-dupes": ["warn", "branch"],
     "jasmine/no-suite-dupes": ["warn", "branch"],
-
     "valid-jsdoc": [
       "warn",
       {
@@ -34,6 +43,5 @@
     "comma-dangle": ["warn", "never"],
     "consistent-return": "off",
     "no-restricted-properties":"off"
-
   }
 }
