+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules
 npm-debug.log
 docs
-lib
+dist
 .eslintcache
 .npmrc
