--- babel.config.cjs.json
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": "defaults",
-        "modules": "commonjs"
-      }
-    ]
-  ],
-  "plugins": ["@babel/plugin-transform-runtime"]
-}
