--- babel.config.json
@@ -1,13 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": {
-          "esmodules": true
-        },
-        "modules": false
-      }
-    ]
-  ]
-}
