--- index.js
@@ -1,4 +1,0 @@
-const axiosRetry = require('./lib/cjs/index').default;
-
-module.exports = axiosRetry;
-module.exports.default = axiosRetry;
