+++ package.json
@@ -12,11 +12,11 @@
     "index.d.ts"
   ],
   "scripts": {
-    "lint": "eslint es/**/*.mjs spec/**/*.spec.mjs",
+    "lint": "eslint src/*.ts",
     "pretest": "npm run lint",
-    "test": "NODE_OPTIONS=--es-module-specifier-resolution=node jasmine",
+    "test": "ts-node --project tsconfig-cjs.json node_modules/jasmine/bin/jasmine",
     "prebuild": "npm run test",
-    "build": "rm -rf lib && babel es -d lib/esm --source-maps && babel es -d lib/cjs --config-file ./babel.config.cjs.json --source-maps && ./fixup",
+    "build": "rm -fr dist/* && tsc -p tsconfig.json && tsc -p tsconfig-cjs.json && ./fixup",
     "prerelease": "npm run build",
     "release": "npm version -m \"New version: %s\"",
     "postrelease": "npm run push && npm publish",
@@ -24,33 +24,32 @@
     "prepare": "husky install"
   },
   "lint-staged": {
-    "*.+(js|mjs)": [
+    "*.ts": [
       "eslint --cache --fix",
       "prettier --write"
-    ],
-    "*.js": "eslint --cache --fix"
+    ]
   },
   "dependencies": {
-    "@babel/runtime": "^7.15.4",
+    "axios": "^1.6.2",
     "is-retry-allowed": "^2.2.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.15.7",
-    "@babel/core": "^7.15.5",
-    "@babel/plugin-transform-runtime": "^7.15.8",
-    "@babel/preset-env": "^7.15.6",
-    "axios": "^1.2.3",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jasmine": "^4.1.2",
-    "eslint-plugin-prettier": "^4.0.0",
-    "husky": "^7.0.2",
-    "jasmine": "^3.9.0",
-    "lint-staged": "^11.2.0",
-    "nock": "^13.1.3",
-    "prettier": "^2.4.1"
+    "@types/axios": "^0.14.0",
+    "@types/node": "^20.9.1",
+    "@types/jasmine": "^5.1.2",
+    "@typescript-eslint/eslint-plugin": "^6.11.0",
+    "@typescript-eslint/parser": "^6.11.0",
+    "eslint": "^8.53.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-jasmine": "^4.1.3",
+    "eslint-plugin-prettier": "^5.0.1",
+    "husky": "^8.0.3",
+    "jasmine": "^5.1.0",
+    "lint-staged": "^15.1.0",
+    "nock": "^13.3.8",
+    "prettier": "^3.1.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
   },
   "repository": {
     "type": "git",
@@ -59,14 +58,13 @@
   "bugs": {
     "url": "https://github.com/softonic/axios-retry/issues"
   },
-  "types": "./index.d.ts",
-  "main": "index.js",
-  "module": "lib/esm/index.js",
+  "types": "dist/esm/index.d.ts",
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
   "exports": {
     ".": {
-      "types": "./index.d.ts",
-      "import": "./lib/esm/index.js",
-      "require": "./index.js"
+      "import": "./dist/esm/index.js",
+      "require": "./dist/cjs/index.js"
     },
     "./package.json": "./package.json"
   }
