+++ spec/support/jasmine.json
@@ -1,6 +1,6 @@
 {
   "spec_dir": "spec",
-  "spec_files": ["**/*[sS]pec.mjs"],
+  "spec_files": ["**/*[sS]pec.ts"],
   "stopSpecOnExpectationFailure": false,
   "random": false
 }
