+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig-base.json",
+  "compilerOptions": {
+      "module": "esnext",
+      "outDir": "dist/esm",
+      "target": "esnext"
+  }
+}
